+++ cypress/e2e/layers.cy.ts
@@ -550,6 +550,104 @@ describe("layers", () => {
     });
   });
 
+  describe("hillshade", () => {
+    it("add", () => {
+      const id = when.modal.fillLayers({
+        type: "hillshade",
+        layer: "example",
+      });
+
+      then(get.styleFromLocalStorage()).shouldDeepNestedInclude({
+        layers: [
+          {
+            id: id,
+            type: "hillshade",
+            source: "example",
+          },
+        ],
+      });
+    });
+
+    it("set hillshade illumination direction array", () => {
+      const id = when.modal.fillLayers({
+        type: "hillshade",
+        layer: "example",
+      });
+      when.collapseGroupInLayerEditor();
+      when.collapseGroupInLayerEditor(1);
+      when.setValueToPropertyArray("spec-field:hillshade-illumination-direction", '1');
+      when.addValueToPropertyArray("spec-field:hillshade-illumination-direction", '2');
+      when.addValueToPropertyArray("spec-field:hillshade-illumination-direction", '3');
+      when.addValueToPropertyArray("spec-field:hillshade-illumination-direction", '4');
+
+      then(get.styleFromLocalStorage()).shouldDeepNestedInclude({
+        layers: [
+          {
+            id: id,
+            type: "hillshade",
+            source: "example",
+            paint: {
+              "hillshade-illumination-direction": [ 1, 2, 3, 4 ]
+            }
+          },
+        ],
+      });
+    });
+
+    it("set hillshade highlight color array", () => {
+      const id = when.modal.fillLayers({
+        type: "hillshade",
+        layer: "example",
+      });
+      when.collapseGroupInLayerEditor();
+      when.setValueToPropertyArray("spec-field:hillshade-highlight-color", 'blue');
+      when.addValueToPropertyArray("spec-field:hillshade-highlight-color", '#00ff00');
+      when.addValueToPropertyArray("spec-field:hillshade-highlight-color", 'rgba(255, 255, 0, 1)');
+
+      then(get.styleFromLocalStorage()).shouldDeepNestedInclude({
+        layers: [
+          {
+            id: id,
+            type: "hillshade",
+            source: "example",
+            paint: {
+              "hillshade-highlight-color": [ "blue", "#00ff00", "rgba(255, 255, 0, 1)" ]
+            }
+          },
+        ],
+      });
+    });
+  });
+
+  describe("color-relief", () => {
+    it("add", () => {
+      const id = when.modal.fillLayers({
+        type: "color-relief",
+        layer: "example",
+      });
+
+      then(get.styleFromLocalStorage()).shouldDeepNestedInclude({
+        layers: [
+          {
+            id: id,
+            type: "color-relief",
+            source: "example",
+          },
+        ],
+      });
+    });
+
+    it("adds elevation expression when clicking the elevation button", () => {
+      when.modal.fillLayers({
+        type: "color-relief",
+        layer: "example",
+      });
+      when.collapseGroupInLayerEditor();
+      when.click("make-elevation-function");
+      then(get.element("[data-wd-key='spec-field-container:color-relief-color'] .CodeMirror-line")).shouldBeVisible();
+    });
+  });
+
   describe("groups", () => {
     it("simple", () => {
       when.setStyle("geojson");