+++ src/components/FieldFunction.tsx
@@ -90,6 +90,18 @@ function getDataType(value: any, fieldSpec={} as any) {
   else if (fieldSpec.type === "array" && isArrayOfPrimatives(value)) {
     return "value";
   }
+  else if (fieldSpec.type === "numberArray" && isArrayOfPrimatives(value)) {
+    return "value";
+  }
+  else if (fieldSpec.type === "colorArray") {
+    return "value";
+  }
+  else if (fieldSpec.type === "padding") {
+    return "value";
+  }
+  else if (fieldSpec.type === "variableAnchorOffsetCollection") {
+    return "value";
+  }
   else if (isZoomField(value)) {
     return "zoom_function";
   }
@@ -293,6 +305,20 @@ const FieldFunction: React.FC<FieldFunctionProps> = (props) => {
     props.onChange(props.fieldName, dataFunc);
   };
 
+  const makeElevationFunction = () => {
+    const expression = [
+      "interpolate",
+      ["linear"],
+      ["elevation"],
+      0,
+      "black",
+      2000,
+      "white"
+    ];
+
+    props.onChange(props.fieldName, expression);
+  };
+
   const onMarkEditing = () => {
     setIsEditing(true);
   };
@@ -364,6 +390,7 @@ const FieldFunction: React.FC<FieldFunctionProps> = (props) => {
         onZoomClick={makeZoomFunction}
         onDataClick={makeDataFunction}
         onExpressionClick={makeExpression}
+        onElevationClick={makeElevationFunction}
       />
     );
   }