+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### ✨ Features and improvements
 - Added translation to "Links" in debug modal
+- Add support for hillshade's color arrays and relief-color elevation expression
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes